src/LocLM/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.25.2
pydantic==2.5.3
ollama==0.1.6
python-dotenv==1.0.0
